requirements.txt
# Core Dependencies
fastmcp==0.2.0
langchain==0.3.18
langchain-openai==0.2.14
langchain-community==0.3.17

# Database
sqlalchemy==2.0.36
sqlite-utils==3.38

# Vector DB & Embeddings
chromadb==1.3.4
sentence-transformers==3.3.1

# Utilities
python-dotenv==1.0.1
pydantic==2.10.5
pydantic-settings==2.7.1

# Development & Testing
pytest==8.3.4
pytest-asyncio==0.25.2
black==24.10.0
ruff==0.8.4
